requirements-model.txt
pandas
pyyaml
scikit-learn
xgboost
mlflow
dvc[s3]